[#3226] - design assignments topbar
