+++ Dashboard/app/js/lib/components/devices/AssignmentsEditView_old/AssignmentDetails.jsx
